+++ packages/cms/lib/modules/choices-guide-widgets/index.js
@@ -49,6 +49,11 @@ module.exports = {
           showFields: ['choicesPreferenceMinColor', 'choicesPreferenceMaxColor','choicesPreferenceTitle','choicesNoPreferenceYetTitle'],
 				},
 				{
+					label: 'Van min naar plus 100',
+					value: 'minus-to-plus-100',
+          showFields: ['choicesPreferenceMinColor', 'choicesPreferenceMaxColor']
+				},
+				{
 					label: 'In een vlak',
 					value: 'plane',
           showFields: ['choicesPreferenceTitle','choicesNoPreferenceYetTitle'],