+++ packages/cms/lib/modules/openstad-pages/lib/load-ideas.js
@@ -21,12 +21,13 @@ module.exports =  function (req, res, next) {
 
   const thisHost = req.headers['x-forwarded-host'] || req.get('host');
   const protocol = req.headers['x-forwarded-proto'] || req.protocol;
-  const siteUrl = protocol + '://' + thisHost;
-  const fullUrl = siteUrl + req.originalUrl;
+  const baseUrl = protocol + '://' + thisHost;
+  const fullUrl = baseUrl + req.originalUrl;
   const parsedUrl = url.parse(fullUrl, true);
 
+
   //add url
-  req.data.siteUrl = siteUrl;
+//  req.data.siteUrl = siteUrl;
   req.data.currentPath = parsedUrl.path;
   req.data.currentPathname = parsedUrl.pathname;
   req.data.currentUrl = fullUrl;
@@ -64,15 +65,17 @@ module.exports =  function (req, res, next) {
        .then(function (ideas) {
          const ideaSlug = req.data.global.ideaSlug;
          const ideaOverviewSlug = req.data.global.ideaOverviewSlug;
-         const protocol = req.headers['x-forwarded-proto'] || req.protocol;
+         const siteUrl = req.data.cmsUrl;
+
+         console.log('siteUrl', siteUrl)
 
          /**
           * Format ideas data
           */
          ideas = ideas.map((idea) => {
            let createdData = new Date(idea.createdAt);
-           idea.fullUrl = `${protocol}://${thisHost}/${ideaSlug}/${idea.id}`;
-           idea.overviewUrl = `${protocol}://${thisHost}/${ideaOverviewSlug}?ideaId=${idea.id}`;
+           idea.fullUrl = `${siteUrl}/${ideaSlug}/${idea.id}`;
+           idea.overviewUrl = `${siteUrl}/${ideaOverviewSlug}?ideaId=${idea.id}`;
            idea.createdTime = createdData.getTime();
 
            if (idea.location && idea.location.coordinates[0] && idea.location.coordinates[1]) {