+++ packages/cms/lib/modules/participatory-budgeting-widgets/public/js/voting.js
@@ -101,7 +101,7 @@ if (votingContainer !== null) {
 
 	  // scrollToBudget()
 
-	  updateBudgetDisplay();
+	  updateBudgetDisplay(false);
 	  updateListElements();
   }
 
@@ -369,7 +369,15 @@ if (votingContainer !== null) {
 	  }
   }
 
-  function updateBudgetDisplay() {
+  function updateBudgetDisplay(productAdded) {
+
+  	// Determine which message to show to the screen reader
+		// Todo: Refactor this into using constants so we have more control
+  	if (typeof productAdded === 'undefined') {
+  		productAdded = true;
+		} else {
+  		productAdded = !!productAdded;
+		}
 
 	  var budgetBarContainer = document.querySelector('#current-budget-bar');
 
@@ -459,7 +467,7 @@ if (votingContainer !== null) {
 			    $('.current-budget-amount').html(formatEuros(initialAvailableBudget - availableBudgetAmount));
 		      $('.available-budget-amount').html(formatEuros(availableBudgetAmount));
 
-		      addCurrentBudgetScreenReaderAlert(initialAvailableBudget, availableBudgetAmount);
+		      addCurrentBudgetScreenReaderAlert(initialAvailableBudget, availableBudgetAmount, productAdded);
 
 			    previewImages.innerHTML = '';
 			    currentSelection.forEach( function(id) {
@@ -471,10 +479,16 @@ if (votingContainer !== null) {
   				    previewImage.setAttribute('data-idea-id', element.ideaId);
   				    previewImage.className += ' idea-' + element.ideaId;
 
-
   				    var linkToIdea = document.createElement("a");
   				    linkToIdea.href = '#ideaId-' + element.ideaId;
+  				    linkToIdea.className = 'idea-preview';
   				    linkToIdea.appendChild(previewImage);
+
+  				    var screenReaderTitle = document.createElement("span");
+  				    screenReaderTitle.className = 'sr-only';
+  				    screenReaderTitle.textContent = 'Plan: ' + element.title + ', Thema: ' + element.getAttribute('data-theme') + ', Budget: € ' + element.getAttribute('data-budget');
+
+  				    linkToIdea.appendChild(screenReaderTitle);
 
   				    previewImages.appendChild(linkToIdea);
             }
@@ -986,10 +1000,10 @@ if (votingContainer !== null) {
   function login() {
 	  logout({
 		  success: function(data) {
-			  window.location.href = '/oauth/login?redirect_uri=' + currentPath;
+			  window.location.href = window.siteUrl + '/oauth/login?redirect_uri=' + currentPath;
 		  },
 		  error: function(error) {
-			  window.location.href = '/oauth/login?redirect_uri='+ currentPath;
+			  window.location.href = window.siteUrl + '/oauth/login?redirect_uri='+ currentPath;
 		  }
 	  });
 
@@ -1482,6 +1496,20 @@ if (votingContainer !== null) {
     }, 1)
   });
 
+
+  $(document).on('click', 'a.idea-preview', function (e) {
+  	var editMode = $(this).closest('#current-budget-preview').hasClass('editMode');
+
+  	if (editMode) {
+  		e.preventDefault();
+  		$(this).find('.idea-image-mask').trigger('click');
+
+  		$(this).closest('#current-budget-preview').find('a.idea-preview:first').focus();
+
+  		return false;
+		}
+	})
+
   function displayIdeaOnHash () {
 
 	  var ideaId;
@@ -1742,16 +1770,30 @@ if (votingContainer !== null) {
   updateBudgetDisplay();
 
   // @todo: create a screen reader alert for count and budgeting-per-theme voting types
-  function addCurrentBudgetScreenReaderAlert (initialAvailableBudget, availableBudgetAmount) {
+  function addCurrentBudgetScreenReaderAlert (initialAvailableBudget, availableBudgetAmount, productAdded) {
 
   	if (votingType != 'budgeting') {
   		return;
 		}
 
+  	if (typeof productAdded === 'undefined') {
+  		productAdded = true;
+		} else {
+  		productAdded = !!productAdded;
+		}
+
+  	// Todo: Refactor this into using constants and a user-changeable text
+  	var planMessage = 'Plan toegevoegd aan winkelmand.';
+
+  	if (!productAdded) {
+			planMessage = 'Plan verwijderd uit winkelmand.';
+		}
+
 		var div = document.createElement('div');
 		div.setAttribute('role', 'alert');
 		div.className = 'sr-only';
-		div.innerHTML = 'Plan toegevoegd aan winkelmand. Uw gekozen budget bedraagt: ' + formatEuros(initialAvailableBudget - availableBudgetAmount) + ' U heeft nog over: ' + formatEuros(availableBudgetAmount);
+
+		div.innerHTML = planMessage +' Uw gekozen budget bedraagt: ' + formatEuros(initialAvailableBudget - availableBudgetAmount) + ' U heeft nog over: ' + formatEuros(availableBudgetAmount);
 		document.body.appendChild(div);
 
 	}